chat-app-backend/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
constantly==15.1.0
coverage==7.4.1
cryptography==40.0.2
daphne==4.0.0
Django==4.2
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
dnspython==2.3.0
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==22.6.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.5
msgpack==1.0.5
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pymongo==4.3.3
pyOpenSSL==23.1.1
pytest==8.0.0
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==4.5.4
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
zope.interface==6.0
